+++ fellow/SRC/WinFellow.Core/Service/IRetroPlatform.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "Defs.h"
+#include <cstdint>
 
-namespace fellow::api::service
+namespace Service
 {
   class IRetroPlatform
   {
