+++ fellow/SRC/WinFellow.Test.Infrastructure/mock/service/RetroPlatformWrapperMock.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "fellow/api/service/IRetroPlatform.h"
+#include "Service/IRetroPlatform.h"
 
 namespace test::mock::fellow::api::service
 {
-  class RetroPlatformWrapperMock : public ::fellow::api::service::IRetroPlatform
+  class RetroPlatformWrapperMock : public ::Service::IRetroPlatform
   {
   public:
     bool SendHardDriveContent(const uint32_t lHardDriveNo, const char *szImageName, const bool bWriteProtected) override;
