+++ fellow/SRC/WinFellow/WinFellow.vcxproj
@@ -487,6 +487,8 @@ PowerShell.exe ..\build\Scripts\GitWCRev.ps1 -InputFileName GUI_versioninfo-wcre
     <ClCompile Include="VirtualHost\CoreFactory.cpp" />
     <ClCompile Include="Windows\Driver\Sound\DirectSoundDriver.cpp" />
     <ClCompile Include="Windows\Service\FileopsWin32.cpp" />
+    <ClCompile Include="Windows\Service\Hud.cpp" />
+    <ClCompile Include="Windows\Service\RetroPlatformWrapper.cpp" />
     <ClCompile Include="zlib\C\gzclose.c">
       <CompileAs Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">CompileAsC</CompileAs>
       <CompileAs Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">CompileAsC</CompileAs>
@@ -1122,9 +1124,6 @@ PowerShell.exe ..\build\Scripts\GitWCRev.ps1 -InputFileName GUI_versioninfo-wcre
     <ClCompile Include="DXGI\GfxDrvDXGIModeEnumerator.cpp" />
     <ClCompile Include="DXGI\GfxDrvDXGIOutput.cpp" />
     <ClCompile Include="DXGI\GfxDrvDXGIOutputEnumerator.cpp" />
-    <ClCompile Include="fellow\service\HUD.cpp" />
-    <ClCompile Include="fellow\service\RetroPlatformWrapper.cpp" />
-    <ClCompile Include="fellow\service\Services.cpp" />
     <ClCompile Include="fellow\vm\M68K.cpp" />
     <ClCompile Include="fellow\vm\MemorySystem.cpp" />
     <ClCompile Include="fellow\vm\VM.cpp" />
@@ -1176,6 +1175,8 @@ PowerShell.exe ..\build\Scripts\GitWCRev.ps1 -InputFileName GUI_versioninfo-wcre
     <ClInclude Include="Windows\Driver\Sound\DirectSoundDriver.h" />
     <ClInclude Include="Windows\Driver\Sound\DirectSoundMode.h" />
     <ClInclude Include="Windows\Service\FileopsWin32.h" />
+    <ClInclude Include="Windows\Service\Hud.h" />
+    <ClInclude Include="Windows\Service\RetroPlatformWrapper.h" />
     <ClInclude Include="xdms\include\cdata.h" />
     <ClInclude Include="xdms\include\crc_csum.h" />
     <ClInclude Include="xdms\include\getbits.h" />
@@ -1209,14 +1210,9 @@ PowerShell.exe ..\build\Scripts\GitWCRev.ps1 -InputFileName GUI_versioninfo-wcre
     <ClInclude Include="DXGI\GfxDrvDXGIModeEnumerator.h" />
     <ClInclude Include="DXGI\GfxDrvDXGIOutput.h" />
     <ClInclude Include="DXGI\GfxDrvDXGIOutputEnumerator.h" />
-    <ClInclude Include="fellow\api\Services.h" />
-    <ClInclude Include="fellow\api\service\IHUD.h" />
-    <ClInclude Include="fellow\api\service\IRetroPlatform.h" />
     <ClInclude Include="fellow\api\VM.h" />
     <ClInclude Include="fellow\api\vm\IM68K.h" />
     <ClInclude Include="fellow\api\vm\IMemorySystem.h" />
-    <ClInclude Include="fellow\service\HUD.h" />
-    <ClInclude Include="fellow\service\RetroPlatformWrapper.h" />
     <ClInclude Include="fellow\vm\M68K.h" />
     <ClInclude Include="fellow\vm\MemorySystem.h" />
     <ClInclude Include="graphics\BitplaneDMA.h" />
