+++ fellow/SRC/WinFellow/Windows/RetroPlatform.h
@@ -35,10 +35,10 @@
 #include "dxver.h"
 #include <dinput.h>
 
+#include "Defs.h"
 #include "gfxdrv.h"
 #include "RetroPlatformGuestIPC.h"
 #include "RetroPlatformIPC.h"
-#include "Defs.h"
 
 #define RETRO_PLATFORM_HARDDRIVE_BLINK_MSECS 100
 #define RETRO_PLATFORM_KEYSET_COUNT 6 ///< north, east, south, west, fire, autofire
